TreeNode: initial port to BoardNode

+++ TreeNode.conf.h
@@ -1,2 +1,2 @@
 // One dimensional tweak space, need to optimize with Yelps MOE
-const float TreeNode::explorationParameter = 0.35;
+const double TreeNode::explorationParameter = 0.35;
